agent-orchestrator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
asyncpg==0.29.0
redis==5.0.4
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pydantic==1.10.15
